Omnijaws: widget: Uncouple the temperature from the image
Until this change, the temperature strings were combined together with the weather condition images. This caused the temperature string to be clipped when using some icon packs.

How it looks: https://imgur.com/a/Kj3w3z3

Let's just uncouple the text and the image and set the text to a separate Textview. Also improves readability

Signed-off-by: Ray <vincentrayananda@gmail.com>
Change-Id: I6b53e7e6ea86c95f4018551579c0a26bd580b874